Reorganize field formatters and adjust namespacing
No code changes, name changes only.
