+++ src/globus_cli/commands/collection/show.py
@@ -21,7 +21,7 @@ def _get_standard_fields(auth_client: globus_sdk.AuthClient) -> list[Field]:
         Field(
             "Owner",
             "identity_id",
-            formatter=formatters.IdentityStrFormatter(auth_client),
+            formatter=formatters.auth.IdentityIDFormatter(auth_client),
         ),
         Field("ID", "id"),
         Field("Collection Type", "collection_type"),
