+++ src/globus_cli/commands/endpoint/permission/_common.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from globus_cli.termio import formatters
 
 
-class AclPrincipalFormatter(formatters.PrincipalDictFormatter):
+class AclPrincipalFormatter(formatters.auth.PrincipalDictFormatter):
     # customize the formatter to provide the `principal_type` as the fallback value for
     # unrecognized types. This handles various cases in which
     # `principal_type=all_authenticated_users` or similar, which is the shape of the
